+++ Tests/LegacyUnit/BagBuilder/EventBagBuilderTest.php
@@ -6298,6 +6298,22 @@ final class EventBagBuilderTest extends TestCase
     /**
      * @test
      */
+    public function limitToOrganizersForOneExistingAndOneInexistentOrganizerFindsEventWithExistingOrganizer(): void
+    {
+        $eventUid = $this->testingFramework->createRecord('tx_seminars_seminars');
+        $organizerUid1 = $this->testingFramework->createRecord('tx_seminars_organizers');
+        $this->testingFramework
+            ->createRelationAndUpdateCounter('tx_seminars_seminars', $eventUid, $organizerUid1, 'organizers');
+
+        $this->subject->limitToOrganizers($organizerUid1 . ',' . ($organizerUid1 + 1));
+        $bag = $this->subject->build();
+
+        self::assertSame(1, $bag->count());
+    }
+
+    /**
+     * @test
+     */
     public function limitToOrganizersForProvidedOrganizerAndTwoEventsWithThisOrganizerFindsTheseEvents(): void
     {
         $eventUid1 = $this->testingFramework->createRecord(
@@ -6398,6 +6414,27 @@ final class EventBagBuilderTest extends TestCase
             1,
             $bag->count()
         );
+    }
+
+    /**
+     * @test
+     *
+     * @param int|non-empty-string $invalidUid
+     *
+     * @dataProvider nonPositiveIntegerDataProvider
+     * @dataProvider sqlCharacterDataProvider
+     */
+    public function limitToOrganizersSilentlyIgnoreInvalidUids($invalidUid): void
+    {
+        $eventUid = $this->testingFramework->createRecord('tx_seminars_seminars');
+        $organizerUid1 = $this->testingFramework->createRecord('tx_seminars_organizers');
+        $this->testingFramework
+            ->createRelationAndUpdateCounter('tx_seminars_seminars', $eventUid, $organizerUid1, 'organizers');
+
+        $this->subject->limitToOrganizers($organizerUid1 . ',' . $invalidUid);
+        $bag = $this->subject->build();
+
+        self::assertSame(1, $bag->count());
     }
 
     ////////////////////////////////